FEATURE: Allows multiple updates per update, files
New set_multi_attribute method allows you to pass a hash with key/value pairs,
all of which are updated in one API call. If a value is a 'file' parameter
(LSHW, LLDP) and the value is prefixed with an '@' sign, the client code will
assume it is a file and attempt to read it.

Via collins-shell you can now also use set_attributes (instead of set_attribute)
with multiple key=value parameters and those will all be processed in one call.
The '@' sign convention follows in the shell as well. This convention comes from
curl.